BaSP-TG-107: Work in progress adding routes
