+++ src/Components/TimeSheets/ListItem.jsx
@@ -1,14 +1,19 @@
 import React from 'react';
 
-const ListItem = ({ listItem, deleteItem, editMode, updateItem, props }) => {
+const ListItem = ({ listItem, deleteItem }) => {
   const handleDelete = (_id) => {
     deleteItem(_id);
   };
-  const handleUpdate = (_id) => {
-    updateItem(_id);
-    editMode();
-    props.history.push('/time-sheets/edit');
+
+  const editTimeSheet = (_id) => {
+    window.location = `/time-sheets/edit?=${_id}`;
   };
+
+  // const handleUpdate = (_id) => {
+  //   updateItem(_id);
+  //   editMode();
+  //   props.history.push('/time-sheets/edit');
+  // };
   const amountOfTasks = (tasks) => {
     if (tasks.length === 1) {
       return tasks[0].description;
@@ -39,7 +44,7 @@ const ListItem = ({ listItem, deleteItem, editMode, updateItem, props }) => {
           <button onClick={() => handleDelete(item._id)}>X</button>
         </td>
         <td>
-          <button onClick={() => handleUpdate(item._id)}>Edit</button>
+          <button onClick={() => editTimeSheet(item._id)}>Edit</button>
         </td>
       </tr>
     );
