+++ src/Components/TimeSheets/TimeSheetsForm.jsx
@@ -1,15 +1,6 @@
 import React, { useState, useEffect } from 'react';
 
-const TimeSheetsForm = ({
-  edit,
-  itemToUpdate,
-  editStartDate,
-  handleEditStartDate,
-  editEndDate,
-  handleEditEndDate,
-  switcher,
-  props
-}) => {
+const TimeSheetsForm = ({ props }) => {
   const [addItem, setItem] = useState({});
   const [employeesItem, setEmployeesItem] = useState([]);
   const [projectsItem, setProjectsItem] = useState([]);
@@ -19,6 +10,46 @@ const TimeSheetsForm = ({
   const [taskList, setTaskList] = useState(emptyList);
   const [employeeNameExist, setEmployeeNameExist] = useState(false);
   const [projectNameExist, setProjectNameExist] = useState(false);
+
+  let [edit, setEdit] = useState(false);
+  const [itemToUpdate, setItemToUpdate] = useState();
+  const [editStartDate, setEditStartDate] = useState(true);
+  const [editEndDate, setEditEndDate] = useState(true);
+
+  const params = window.location.search;
+  let idParam = params.substring(2);
+
+  if (idParam != 0) {
+    console.log('id', idParam);
+    updateItem(idParam);
+    editMode();
+  }
+
+  useEffect(() => {
+    setEditStartDate(true);
+    setEditEndDate(true);
+  }, []);
+
+  const updateItem = (id) => {
+    setItemToUpdate(list.filter((timeSheet) => timeSheet._id === id));
+  };
+
+  if (edit) {
+    setEdit(edit ? (edit = false) : (edit = true));
+    handleEditStartDate;
+  }
+
+  const handleEditStartDate = (state) => {
+    setEditStartDate(state);
+  };
+
+  const handleEditEndDate = (state) => {
+    setEditEndDate(state);
+  };
+
+  const editMode = () => {
+    setEdit(edit ? (edit = false) : (edit = true));
+  };
 
   useEffect(() => {
     if (edit) {
@@ -109,7 +140,7 @@ const TimeSheetsForm = ({
             .then((res) => {
               alert(res.msg);
               if (!res.error) {
-                switcher();
+                props.history.push('/time-sheets');
               }
             });
         } catch (error) {
@@ -127,9 +158,8 @@ const TimeSheetsForm = ({
           .then((response) => {
             alert(response.error ? `Error! ${response.msg}` : `Success! ${response.message}`);
             if (!response.error) {
-              switcher();
+              props.history.push('/time-sheets');
             }
-            props.history.push('/time-sheets');
           });
       } catch (error) {
         console.error(error);
@@ -162,6 +192,8 @@ const TimeSheetsForm = ({
       </div>
       <form onSubmit={create}>
         <div>
+          <h2>TimeSheets</h2>
+          <button>back</button>
           <label>Employee</label>
           <select onChange={onChange} name="employeeId">
             {
