+++ src/Components/TimeSheets/index.js
@@ -1,15 +1,9 @@
 import styles from './time-sheets.module.css';
 import TableList from './TableList';
 import React, { useState, useEffect } from 'react';
-import TimeSheetsForm from './TimeSheetsForm';
 
 function TimeSheets(props) {
-  let [change, setSwitch] = useState(false);
-  let [edit, setEdit] = useState(false);
   const [list, setList] = useState([]);
-  const [itemToUpdate, setItemToUpdate] = useState();
-  const [editStartDate, setEditStartDate] = useState(true);
-  const [editEndDate, setEditEndDate] = useState(true);
 
   useEffect(() => {
     fetch(`https://coco-trackgenix-server.vercel.app/timesheets`)
@@ -18,11 +12,6 @@ function TimeSheets(props) {
         setList(json.data);
       });
   }, []);
-
-  useEffect(() => {
-    setEditStartDate(true);
-    setEditEndDate(true);
-  }, [change]);
 
   const deleteItem = (id) => {
     setList([...list.filter((listItem) => listItem._id !== id)]);
@@ -37,64 +26,14 @@ function TimeSheets(props) {
     }
   };
 
-  const updateItem = (id) => {
-    setItemToUpdate(list.filter((timeSheet) => timeSheet._id === id));
-  };
-
-  const switcher = () => {
-    setSwitch(change ? (change = false) : (change = true));
-    props.history.push('/time-sheets/add');
-    if (edit) {
-      setEdit(edit ? (edit = false) : (edit = true));
-      handleEditStartDate;
-    }
-  };
-
-  const handleEditStartDate = (state) => {
-    setEditStartDate(state);
-  };
-
-  const handleEditEndDate = (state) => {
-    setEditEndDate(state);
-  };
-
-  const editMode = () => {
-    setEdit(edit ? (edit = false) : (edit = true));
-    setSwitch(change ? (change = false) : (change = true));
-  };
-
-  if (change) {
-    return (
-      <section className={styles.container}>
-        <h2>TimeSheets</h2>
-        <button onClick={switcher}>back</button>
-        <TimeSheetsForm
-          edit={edit}
-          itemToUpdate={itemToUpdate}
-          editStartDate={editStartDate}
-          editEndDate={editEndDate}
-          handleEditStartDate={handleEditStartDate}
-          handleEditEndDate={handleEditEndDate}
-          switcher={switcher}
-        />
-      </section>
-    );
-  } else {
-    return (
-      <section className={styles.container}>
-        <h2>TimeSheets</h2>
-        <button className={styles.addButton} onClick={() => props.history.push('/time-sheets/add')}>
-          Add new
-        </button>
-        <TableList
-          list={list}
-          setList={setList}
-          deleteItem={deleteItem}
-          editMode={editMode}
-          updateItem={updateItem}
-        />
-      </section>
-    );
-  }
+  return (
+    <section className={styles.container}>
+      <h2>TimeSheets</h2>
+      <button className={styles.addButton} onClick={() => props.history.push('/time-sheets/add')}>
+        Add new
+      </button>
+      <TableList list={list} setList={setList} deleteItem={deleteItem} />
+    </section>
+  );
 }
 export default TimeSheets;
